all cosmetics and project clean up
ready for api testing
