+++ orders-pick-and-pack/src/app/pages/logistics/logistics-detail-modal/logistics-detail.component.ts
@@ -1,6 +1,6 @@
 import {Component, ElementRef, Inject, OnInit, ViewChild} from '@angular/core';
 import {FormBuilder, FormGroup, Validators} from '@angular/forms';
-import {OrderService} from "../../../@core/services/order.service";
+import {LogisticsService} from "../../../@core/services/logistics.service";
 import {OrderModel} from "../../../@core/models/order.model";
 import {ClickEventArgs} from "@syncfusion/ej2-navigations";
 
@@ -11,7 +11,8 @@ import {LogisticsStatusService} from "../../../@core/services/logistics.status.s
 import {takeUntil} from "rxjs/operators";
 import {ActivatedRoute} from "@angular/router";
 import {Subject} from "rxjs";
-import {NbDialogRef} from "@nebular/theme";
+import {NbDialogRef, NbToastrService} from "@nebular/theme";
+import {ToastService} from "../../../@core/services/toast.service";
 
 @Component({
     selector: 'logistics-detail-modal',
@@ -21,8 +22,23 @@ import {NbDialogRef} from "@nebular/theme";
 export class LogisticsDetailComponent implements OnInit {
     // public logisticsDetailForm: FormGroup;
 
-    formValidators = [Validators.required];
+    private _destroy$: Subject<void> = new Subject<void>();
+    editMode: boolean = false;
+
+    formValidators = [Validators.required,];
     logisticsDetailForm: FormGroup = new FormGroup({});
+    dialogTitle: string = "View";
+
+    toggleEdit() {
+        this.editMode = !this.editMode;
+        if (this.editMode) {
+            this.dialogTitle = "Update";
+            this.logisticsDetailForm.enable();
+        } else {
+            this.dialogTitle = "Update";
+            this.logisticsDetailForm.disable();
+        }
+    }
 
     // scan modal
     // @ViewChild('scanItemModal') scanItemModal: DialogComponent | any;
@@ -78,9 +94,10 @@ export class LogisticsDetailComponent implements OnInit {
 
     constructor(private formBuilder: FormBuilder,
                 private route: ActivatedRoute,
+                private toasterService: NbToastrService,
                 private nbDialogRef: NbDialogRef<LogisticsDetailComponent>,
                 private logisticStatusService: LogisticsStatusService,
-                private orderService: OrderService) {
+                private orderService: LogisticsService) {
         this.orderService.currentSelectedOrderObservable.subscribe(
             {
                 next: logisticModel => {
@@ -92,15 +109,34 @@ export class LogisticsDetailComponent implements OnInit {
 
     initializeForm() {
         this.logisticsDetailForm = this.formBuilder.group({
-            OrderNumber: [this.logisticsModel.Order.OrderNumber, [...this.formValidators]],
-            Id: [this.logisticsModel.Id, [...this.formValidators]],
-            LogisticsStatusId: [this.logisticsModel.LogisticsStatusId, [...this.formValidators]],
-            DistributionId: [this.logisticsModel.DistributionId, [...this.formValidators]],
-            CollectionId: [this.logisticsModel.CollectionId, [...this.formValidators]],
-            UpdateBy: [this.logisticsModel.UpdateBy, [...this.formValidators]],
-            DateCreated: [this.logisticsModel.CreatedDate, [...this.formValidators]],
-            TotalItems: [this.logisticsModel.Order.OrderItems.length, [...this.formValidators]],
-            UpdateDate: [this.logisticsModel.UpdateDate, [...this.formValidators]],
+            OrderNumber: [{
+                value: this.logisticsModel.Order.OrderNumber,
+                disabled: !this.editMode
+            }, [...this.formValidators]],
+
+            Id: [{value: this.logisticsModel.Id, disabled: !this.editMode}, [...this.formValidators]],
+            LogisticsStatusId: [{
+                value: this.logisticsModel.LogisticsStatusId,
+                disabled: !this.editMode
+            }, [...this.formValidators]],
+
+            DistributionId: [{
+                value: this.logisticsModel.DistributionId,
+                disabled: !this.editMode
+            }, [...this.formValidators]],
+
+            CollectionId: [{
+                value: this.logisticsModel.CollectionId,
+                disabled: !this.editMode
+            }, [...this.formValidators]],
+
+            UpdateBy: [{value: this.logisticsModel.UpdateBy, disabled: !this.editMode}, [...this.formValidators]],
+            DateCreated: [{value: this.logisticsModel.CreatedDate, disabled: !this.editMode}, [...this.formValidators]],
+            TotalItems: [{
+                value: this.logisticsModel.Order.OrderItems.length,
+                disabled: !this.editMode
+            }, [...this.formValidators]],
+            UpdateDate: [{value: this.logisticsModel.UpdateDate, disabled: !this.editMode}, [...this.formValidators]],
 
         });
     }
@@ -117,7 +153,6 @@ export class LogisticsDetailComponent implements OnInit {
         })
     }
 
-    private _destroy$: Subject<void> = new Subject<void>();
 
     setScope() {
         this.route.queryParams
@@ -174,5 +209,14 @@ export class LogisticsDetailComponent implements OnInit {
 
     closeModal() {
         this.nbDialogRef.close();
+    }
+
+    submitChanges() {
+        if (this.logisticsDetailForm.valid) {
+            this.orderService.updateLogistic(this.logisticsDetailForm.value);
+        } else {
+            this.toasterService.warning("You have some errors in your form","Form Not Complete");
+        }
+
     }
 }
